refactor(digitalcombo): Rename lcec_el1859 to lcec_digtalcombo
