+++ documentation/devices/EL1852.yml
@@ -7,5 +7,5 @@ Description: Beckhoff EL1852 8Ch. Dig. Input 24V, 3ms, 8Ch. Dig. Output 24V, 0.5
 DocumentationURL: http://www.beckhoff.com/EL1852
 DeviceType: Digital Input/Output
 Notes: ""
-SrcFile: src/devices/lcec_el1859.c
+SrcFile: src/devices/lcec_digitalcombo.c
 TestingStatus: ""
