+++ documentation/devices/EP2339.yml
@@ -7,5 +7,5 @@ Description: Beckhoff EP2339 16 Ch. Dig. Input/Output 24V, 0,5A, 10-Pole Plug Co
 DocumentationURL: https://www.beckhoff.com/EP2339-0003
 DeviceType: Digital Input/Output
 Notes: "NOTE: some r16 EP2339 devices have a different PDO layout and will not work."
-SrcFile: src/devices/lcec_el1859.c
+SrcFile: src/devices/lcec_digitalcombo.c
 TestingStatus: ""
