+++ documentation/devices/EPP2316.yml
@@ -7,5 +7,5 @@ Description: Beckhoff EP2316-0003 8 Ch. Dig. In, 10µs, 8Ch. Dig. Out 24V, 0,5A,
 DocumentationURL: https://www.beckhoff.com/EP2316-0003
 DeviceType: EtherCAT Fieldbus Boxes (EP23xx)
 Notes: ""
-SrcFile: devices/lcec_el1859.c
+SrcFile: devices/lcec_digitalcombo.c
 TestingStatus: ""
