+++ documentation/devices/EPP2318.yml
@@ -7,5 +7,5 @@ Description: Beckhoff EPP2318-0001 4 Ch. Dig. In, 10µs, 4 Ch. Dig. Out 24V, 0,5
 DocumentationURL: https://www.beckhoff.com/EPP2318-0001
 DeviceType: Digital Input/Output
 Notes: ""
-SrcFile: src/devices/lcec_el1859.c
+SrcFile: src/devices/lcec_digitalcombo.c
 TestingStatus: ""
