+++ documentation/devices/EPP2328.yml
@@ -7,5 +7,5 @@ Description: Beckhoff EPP2328-0001 4 CH. Dig. In, 3ms, 4 Ch. Dig. Out 24V, 2A, M
 DocumentationURL: https://www.beckhoff.com/EPP2328-0001
 DeviceType: Digital Input/Output
 Notes: ""
-SrcFile: src/devices/lcec_el1859.c
+SrcFile: src/devices/lcec_digitalcombo.c
 TestingStatus: ""
