+++ documentation/devices/EPP2334.yml
@@ -7,5 +7,5 @@ Description: Beckhoff EPP2334-0061 4 Ch. Dig. Input/Output 24V, 0,5A, M8
 DocumentationURL: https://www.beckhoff.com/EPP2334-0061
 DeviceType: Digital Input/Output
 Notes: ""
-SrcFile: src/devices/lcec_el1859.c
+SrcFile: src/devices/lcec_digitalcombo.c
 TestingStatus: ""
