+++ documentation/devices/EPP2339.yml
@@ -7,5 +7,5 @@ Description: Beckhoff EPP2339-0003 16 Ch. Dig. Input/Output 24V, 0,5A, 10-Pole P
 DocumentationURL: https://www.beckhoff.com/EPP2339-0003
 DeviceType: Digital Input/Output
 Notes: ""
-SrcFile: src/devices/lcec_el1859.c
+SrcFile: src/devices/lcec_digitalcombo.c
 TestingStatus: ""
